Ticket #1692: force initialisation of empty arrays with zeros and do explicit conversion

+++ src/Common/helpers_platform_dep/freertos/p_sopc_date_time.c
@@ -72,7 +72,7 @@ SOPC_DateTime SOPC_Time_GetCurrentTimeUTC(void)
         return INT64_MAX;
     }
 
-    datetime += currentTimeFrac100NS;
+    datetime += (int64_t) currentTimeFrac100NS;
     return datetime;
 }
